Remove note about adding other DB backends.
libdb5.3 is already present in the -slim images, but not linked, hence
we aren't copying it over. As far as I can tell, it's needed to build
and use the berkeleydb package from PyPI, so (at least in theory) can be
installed and copied over from a venv/builder stage.